move NbsConfiguration from questionbank to database entities module
