+++ apps/question-bank/build.gradle
@@ -96,6 +96,7 @@ dependencies {
     implementation project(':authentication')
     implementation project(':accumulation')
     implementation project(':audit')
+    implementation(project(':database-entities'))
     implementation project(':id-generator')
     implementation(project(':web'))
     implementation(project(':classic-interaction'))
