+++ libs/database-entities/src/main/java/gov/cdc/nbs/repository/NbsConfigurationRepository.java
@@ -1,6 +1,6 @@
-package gov.cdc.nbs.questionbank.question.repository;
+package gov.cdc.nbs.repository;
 
-import gov.cdc.nbs.questionbank.entity.NbsConfiguration;
+import gov.cdc.nbs.entity.odse.NbsConfiguration;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface NbsConfigurationRepository extends JpaRepository<NbsConfiguration, String> {}
